requirements.txt
aiohttp==3.7.4.post0
astroid==2.8.0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cffi==1.14.6
chardet==4.0.0
discord==2.3.2
discord.py==2.3.2
ffmpeg==1.4
idna==3.2
idna-ssl==1.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
platformdirs==2.4.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.11.1
PyNaCl==1.4.0
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
typed_ast==1.4.3
typing-extensions==3.10.0.2
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
